subst. init 3

+++ server/src/controllers/appreciation.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express'
 import { appreciationService } from '../services/appreciation.service.js'
+import { UnauthorizedError } from '../utils/errors.js'
 
 /**
  * Appreciation controller - handles HTTP requests/responses
@@ -12,16 +13,22 @@ export const appreciationController = {
   },
 
   async create(req: Request, res: Response) {
-    // TODO: Get userId from auth middleware
-    const userId = 'temp-user-id'
+    const userId = (req as any).userId
+    if (!userId) {
+      throw new UnauthorizedError('Authentication required')
+    }
+
     const { writingId } = req.params
     const appreciation = await appreciationService.create(writingId, userId)
     res.status(201).json({ data: appreciation })
   },
 
   async remove(req: Request, res: Response) {
-    // TODO: Get userId from auth middleware
-    const userId = 'temp-user-id'
+    const userId = (req as any).userId
+    if (!userId) {
+      throw new UnauthorizedError('Authentication required')
+    }
+
     const { writingId } = req.params
     await appreciationService.remove(writingId, userId)
     res.status(204).send()
